pyproject.toml
[tool.poetry]
name = "odin-bot-exchanges"
version = "0.1.0"
description = ""
authors = ["adolfrodeno <amvillalobos@uc.cl>"]

[tool.poetry.dependencies]
python = "^3.8"
orionx-python-client = "^0.1.2"
odin-bot-entities = "^0.1.3"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
